requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
coverage==7.4.1
cryptography==42.0.2
defusedxml==0.8.0rc2
Django==4.2.3
django-allauth==0.61.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-oauth-toolkit==2.3.0
django-rest-framework-social-oauth2==1.2.0
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
django-simple-history==3.4.0
djangorestframework==3.14.0
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==23.2.1
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jwcrypto==1.5.3
MarkupSafe==2.1.5
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.2
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
simplejson==3.19.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
